Implement a better widget for translatable fields
This commit adds widgets.TranslatableCharField and widgets.TranslatableTextField
which are MultiWidgets which allow entering text, as well as a language code. This is then
the default widget for forms.TranslatableCharField if no other widget (e.g. TextInput)
is specified. It also adds a ModelAdminMixin which you must use to enable these form fields
on the admin, because the admin is stupid and doesn't use the stuff returned from formfield().

If you use a TextInput widget, the language code comes from the one set on the form field or
settings.LANGUAGE_CODE.

Note! This commit means if you are display translatable fields on forms, you might want to override
the widget to be a TextInput to keep everything working the same.